+++ www/commit-hash.js
@@ -1,1 +1,1 @@
-window.__COMMIT__ = 'a4b36ed6ea208d0aaacbf245bfe15c5947ffc1d9';
+window.__COMMIT__ = '64676ab7628aa19f9ff1e48fd322193454c1a931';